move url related utils into userUrlHandling.ts
