+++ src/util/userUrlHandling.ts
@@ -1,9 +1,11 @@
 import { isString } from "lodash";
+import queryString from "query-string";
 
 import {
     URL_PARAM_KEY_FILE_NAME,
     URL_PARAM_KEY_USER_URL,
     USER_TRAJ_REDIRECTS,
+    URL_PARAM_KEY_TIME,
 } from "../constants";
 
 const googleDriveUrlRegEx =
@@ -97,4 +99,41 @@ export const isOnlineTrajectory = (url: string) => {
         url.includes(URL_PARAM_KEY_USER_URL) ||
         url.includes(URL_PARAM_KEY_FILE_NAME)
     );
+};
+
+export const hasUrlParamsSettings = () => {
+    const urlSettings = [URL_PARAM_KEY_TIME];
+    const parsed = queryString.parse(location.search);
+    for (let index = 0; index < urlSettings.length; index++) {
+        const paramName = urlSettings[index];
+        if (parsed[paramName] !== undefined) {
+            return true;
+        }
+    }
+    return false;
+};
+
+export const editUrlParams = (
+    url: string,
+    value: string,
+    paramKey: string
+): string => {
+    // this returns a new url as a string, does not change current url shown in browser
+    const currentUrl = new URL(url);
+    const params = new URLSearchParams(currentUrl.search);
+    params.set(paramKey, value);
+    currentUrl.search = params.toString();
+    return currentUrl.toString();
+};
+
+export const clearBrowserUrlParams = () => {
+    // Removes the query string from the current URL shown in the browser
+    // ex) https://mysite.com/path?city=seattle -> https://mysite.com/path
+    history.replaceState({}, "", `${location.origin}${location.pathname}`);
+};
+
+export const getUrlParamValue = (url: string, param: string) => {
+    const urlObj = new URL(url);
+    const params = new URLSearchParams(urlObj.search);
+    return params.get(param);
 };
